Conflicts:
	src/im_mad/remotes/common.d/collectd-client_control.sh
	src/onedb/4.2.0_to_4.3.80.rb

+++ include/VirtualMachine.h
@@ -1082,7 +1082,7 @@ public:
     /**
      *  Return the total SIZE of volatile disks
      */
-    static float get_volatile_disk_size(Template * tmpl);
+    static long long get_volatile_disk_size(Template * tmpl);
 
     // ------------------------------------------------------------------------
     // Context related functions
